modularise linalg ops
