+++ tfjs-core/src/ops/qr.ts
@@ -1,6 +1,6 @@
 /**
  * @license
- * Copyright 2018 Google LLC. All Rights Reserved.
+ * Copyright 2020 Google Inc. All Rights Reserved.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -14,199 +14,13 @@
  * limitations under the License.
  * =============================================================================
  */
-
-/**
- * Linear algebra ops.
- */
-
 import {ENGINE} from '../engine';
 import {dispose} from '../globals';
-import {Tensor, Tensor1D, Tensor2D} from '../tensor';
-import {convertToTensor} from '../tensor_util_env';
-import {TensorLike} from '../types';
-import {assert} from '../util';
-
-import {squeeze, stack, unstack} from './array_ops';
+import {Tensor, Tensor2D} from '../tensor';
+import {stack, unstack} from './array_ops';
 import {eye} from './eye';
-import {logicalAnd, where} from './logical_ops';
-import {norm} from './norm';
 import {op} from './operation';
-import {sum} from './reduction_ops';
-import {split} from './split';
-import {sub} from './sub';
-import {range, scalar, tensor2d, zeros} from './tensor_ops';
-
-/**
- * Copy a tensor setting everything outside a central band in each innermost
- * matrix to zero.
- *
- * The band part is computed as follows: Assume input has `k` dimensions
- * `[I, J, K, ..., M, N]`, then the output is a tensor with the same shape where
- * `band[i, j, k, ..., m, n] = in_band(m, n) * input[i, j, k, ..., m, n]`.
- * The indicator function
- * `in_band(m, n) = (num_lower < 0 || (m-n) <= num_lower))`
- * `&& (num_upper < 0 || (n-m) <= num_upper)`
- *
- * ```js
- * const x = tf.tensor2d([[ 0,  1,  2, 3],
- *                        [-1,  0,  1, 2],
- *                        [-2, -1,  0, 1],
- *                        [-3, -2, -1, 0]]);
- * let y = tf.linalg.bandPart(x, 1, -1);
- * y.print(); // [[ 0,  1,  2, 3],
- *            //  [-1,  0,  1, 2],
- *            //  [ 0, -1,  0, 1],
- *            //  [ 0, 0 , -1, 0]]
- * let z = tf.linalg.bandPart(x, 2, 1);
- * z.print(); // [[ 0,  1,  0, 0],
- *            //  [-1,  0,  1, 0],
- *            //  [-2, -1,  0, 1],
- *            //  [ 0, -2, -1, 0]]
- * ```
- *
- * @param x Rank `k` tensor
- * @param numLower Number of subdiagonals to keep.
- *   If negative, keep entire lower triangle.
- * @param numUpper Number of subdiagonals to keep.
- *   If negative, keep entire upper triangle.
- * @returns Rank `k` tensor of the same shape as input.
- *   The extracted banded tensor.
- */
-/**
- * @doc {heading:'Operations',
- *       subheading:'Linear Algebra',
- *       namespace:'linalg'}
- */
-function bandPart_<T extends Tensor>(
-    a: T|TensorLike, numLower: number, numUpper: number): T {
-  if (numLower % 1 !== 0) {
-    throw new Error(
-        `bandPart(): numLower must be an integer, got ${numLower}.`);
-  }
-  if (numUpper % 1 !== 0) {
-    throw new Error(
-        `bandPart(): numUpper must be an integer, got ${numUpper}.`);
-  }
-
-  const $a = convertToTensor(a, 'a', 'bandPart');
-
-  if ($a.rank < 2) {
-    throw new Error(`bandPart(): Rank must be at least 2, got ${$a.rank}.`);
-  }
-
-  const shape = $a.shape, [M, N] = $a.shape.slice(-2);
-
-  if (!(numLower <= M)) {
-    throw new Error(
-        `bandPart(): numLower (${numLower})` +
-        ` must not be greater than the number of rows (${M}).`);
-  }
-  if (!(numUpper <= N)) {
-    throw new Error(
-        `bandPart(): numUpper (${numUpper})` +
-        ` must not be greater than the number of columns (${N}).`);
-  }
-
-  if (numLower < 0) {
-    numLower = M;
-  }
-  if (numUpper < 0) {
-    numUpper = N;
-  }
-
-  const i = range(0, M, 1, 'int32').reshape([-1, 1]),
-        j = range(0, N, 1, 'int32'), ij = sub(i, j);
-
-  const inBand = logicalAnd(
-      ij.lessEqual(scalar(+numLower, 'int32')),
-      ij.greaterEqual(scalar(-numUpper, 'int32')));
-
-  const zero = zeros([M, N], $a.dtype);
-
-  return stack(unstack($a.reshape([-1, M, N]))
-                   .map(mat => where(inBand, mat, zero)))
-             .reshape(shape) as T;
-}
-
-/**
- * Gram-Schmidt orthogonalization.
- *
- * ```js
- * const x = tf.tensor2d([[1, 2], [3, 4]]);
- * let y = tf.linalg.gramSchmidt(x);
- * y.print();
- * console.log('Othogonalized:');
- * y.dot(y.transpose()).print();  // should be nearly the identity matrix.
- * console.log('First row direction maintained:');
- * const data = await y.array();
- * console.log(data[0][1] / data[0][0]);  // should be nearly 2.
- * ```
- *
- * @param xs The vectors to be orthogonalized, in one of the two following
- *   formats:
- *   - An Array of `tf.Tensor1D`.
- *   - A `tf.Tensor2D`, i.e., a matrix, in which case the vectors are the rows
- *     of `xs`.
- *   In each case, all the vectors must have the same length and the length
- *   must be greater than or equal to the number of vectors.
- * @returns The orthogonalized and normalized vectors or matrix.
- *   Orthogonalization means that the vectors or the rows of the matrix
- *   are orthogonal (zero inner products). Normalization means that each
- *   vector or each row of the matrix has an L2 norm that equals `1`.
- */
-/**
- * @doc {heading:'Operations',
- *       subheading:'Linear Algebra',
- *       namespace:'linalg'}
- */
-function gramSchmidt_(xs: Tensor1D[]|Tensor2D): Tensor1D[]|Tensor2D {
-  let inputIsTensor2D: boolean;
-  if (Array.isArray(xs)) {
-    inputIsTensor2D = false;
-    assert(
-        xs != null && xs.length > 0,
-        () => 'Gram-Schmidt process: input must not be null, undefined, or ' +
-            'empty');
-    const dim = xs[0].shape[0];
-    for (let i = 1; i < xs.length; ++i) {
-      assert(
-          xs[i].shape[0] === dim,
-          () =>
-              'Gram-Schmidt: Non-unique lengths found in the input vectors: ' +
-              `(${(xs as Tensor1D[])[i].shape[0]} vs. ${dim})`);
-    }
-  } else {
-    inputIsTensor2D = true;
-    xs = split(xs, xs.shape[0], 0).map(x => squeeze(x, [0]));
-  }
-
-  assert(
-      xs.length <= xs[0].shape[0],
-      () => `Gram-Schmidt: Number of vectors (${
-                (xs as Tensor1D[]).length}) exceeds ` +
-          `number of dimensions (${(xs as Tensor1D[])[0].shape[0]}).`);
-
-  const ys: Tensor1D[] = [];
-  const xs1d = xs;
-  for (let i = 0; i < xs.length; ++i) {
-    ys.push(ENGINE.tidy(() => {
-      let x = xs1d[i];
-      if (i > 0) {
-        for (let j = 0; j < i; ++j) {
-          const proj = sum(ys[j].mul(x)).mul(ys[j]);
-          x = x.sub(proj);
-        }
-      }
-      return x.div(norm(x, 'euclidean'));
-    }));
-  }
-
-  if (inputIsTensor2D) {
-    return stack(ys, 0) as Tensor2D;
-  } else {
-    return ys;
-  }
-}
+import {tensor2d} from './tensor_ops';
 
 /**
  * Compute QR decomposition of m-by-n matrix using Householder transformation.
@@ -363,6 +177,4 @@ function qr2d(x: Tensor2D, fullMatrices = false): [Tensor2D, Tensor2D] {
   }) as [Tensor2D, Tensor2D];
 }
 
-export const bandPart = op({bandPart_});
-export const gramSchmidt = op({gramSchmidt_});
 export const qr = op({qr_});
