+++ tfjs-core/src/tests.ts
@@ -45,6 +45,7 @@ import './ops/array_ops_test';
 import './ops/avg_pool_3d_test';
 import './ops/avg_pool_test';
 import './ops/axis_util_test';
+import './ops/band_part_test';
 import './ops/batch_to_space_nd_test';
 import './ops/batchnorm_test';
 import './ops/binary_ops_test';
@@ -75,6 +76,7 @@ import './ops/equal_test';
 import './ops/eye_test';
 import './ops/fused_test';
 import './ops/gather_nd_test';
+import './ops/gram_schmidt_test';
 import './ops/greater_equal_test';
 import './ops/greater_test';
 import './ops/image_ops_test';
@@ -82,7 +84,6 @@ import './ops/in_top_k_test';
 import './ops/leaky_relu_test';
 import './ops/less_equal_test';
 import './ops/less_test';
-import './ops/linalg_ops_test';
 import './ops/local_response_normalization_test';
 import './ops/logical_ops_test';
 import './ops/loss_ops_test';
@@ -98,6 +99,7 @@ import './ops/one_hot_test';
 import './ops/operation_test';
 import './ops/pad_test';
 import './ops/pool_test';
+import './ops/qr_test';
 import './ops/rand_test';
 import './ops/random_gamma_test';
 import './ops/random_normal_test';
